Revert "[GRO-40] - Add amplitude to app (#15)"
This reverts commit 66f9237b1fb9d719c64d7b288532ae0b84a6f40c.
