--- src/components/Providers.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import { usePageSegmentAnalytics } from '@/hooks/use-page-segment-analytics';
-import { PropsWithChildren } from 'react';
-
-export function Providers({ children }: PropsWithChildren) {
-  usePageSegmentAnalytics();
-  return <>{children}</>;
-}
