--- src/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './use-page-segment-analytics';
