--- src/utils/misc/isDefined.ts
@@ -1,5 +1,0 @@
-export const isDefined = <T extends unknown>(
-  value: T
-): value is NonNullable<T> => {
-  return value !== null && value !== undefined;
-};
